Move core logic into rupa/z driver
The default driver of z.vim has been rupa/z for a long time, and more
recently z.lua. Those are currently combined. It'll take some effort to
split 'em out. There's a high likelyhood that this broke everything. But
hey, it works for zoxide...

Yeah I need to fix this.