Remove unit test abominations
I'm guessing a lot of the test hacks were from a time when I didn't
understand vim's "module system" and was naively trying to reach 100%
coverage.

After my experience with Vader, I think 100% coverage is complete
insanity. It's nice when it's nice and horrible when it isn't.

I've ripped out the tests that didn't provide much value. Some are still
failing. That's okay.