Automatically detect support for rupa/z
This implements automatic support detection for the `rupa/z` driver. It
checks the `$ZSH` variable (as it's done historically) or checks the
file name of the configured executable (a new behavior). The file name
check _should_ be consistent. Both omz and the source code refer to the
executable as `z.sh`.