Document z.lua support as new version
I've got a hefty PR incoming, but before I merge it, I'm attaching git
tags to the older releases so people can easily roll back to a version
number without digging through history. The last major event was adding
support for z.lua, so I've drafted a new version in the "changelog" and
I'll be attaching a corresponding git tag.