Add deprecation notice to `zcd#path`
This is an effort to migrate everything to a consistent name, part of
renaming the plugin to `teleport.vim`. If the user specifies `zcd#path`,
they'll get an error and be prompted to rename to `teleport#path`.
I hacked around with some ideas to fail more gracefully, but it would've
required a *lot* more effort on my part. I'm hoping my users will
forgive me.