Polyfill the `trim(...)` function
See pipeline failure:
https://github.com/PsychoLlama/teleport.vim/runs/1520737904?check_suite_focus=true

Vim 8.0 doesn't seem to support the `trim(...)` function. I know it's
supported in 8.2, but unfortunately I know of some users who are still
on v7 series. Poor souls.

Anyway, this implements a polyfill for the `trim(...)` function. It
ain't perfect, but it works well enough for my purposes.